Reorganize directory structure for best practices
Directory layout:
  src/                    # Python source modules
    __init__.py
    cgm_args.py
    Defaults.py
    dexcom_data.py
    http_general.py
    logger.py
    nightscout_data.py
    pygame_display.py
  assets/                 # Images and resources
    images/               # Documentation images
      IMG_0440.jpeg
      IMG_2247.jpeg
      IMG_5750.png
    loop-status/          # Loop status indicator images
      loop-aging@38mm.png
      loop-fresh@38mm.png
      loop-stale@38mm.png
    nightscout_large.png  # Nightscout icon (downloaded at runtime)
  docs/                   # Documentation
    THIRD_PARTY_NOTICES.md
  archive/                # Archived/unused code
    cgm_display_*.py
  cgm_display.py          # Main entry point (root)
  Dockerfile
  README.md
  requirements.txt

Changes:
- Moved Python modules to src/ package with relative imports
- Organized images into assets/images/ and assets/loop-status/
- Moved documentation to docs/ directory
- Updated all imports to use src.* package references
- Created src/__init__.py to establish package structure
