+++ docs/THIRD_PARTY_NOTICES.md
